requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.1
psycopg2==2.9.1